Update core to export Int64
